# docs/swagger.yaml
basePath: /
definitions:
  github_com_fernandovalenzuela_api-home-pay_internal_models.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_models.Company:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      website_url:
        type: string
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_models.Expense:
    properties:
      account_id:
        type: integer
      amount_paid:
        type: number
      category:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Category'
      category_id:
        type: integer
      created_at:
        type: string
      current_amount:
        type: number
      current_installment:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      installment_group_id:
        type: string
      is_recurring:
        type: boolean
      notes:
        type: string
      period:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Period'
      period_id:
        type: integer
      service_account:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
      total_installments:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_fernandovalenzuela_api-home-pay_internal_models.Period:
    properties:
      id:
        type: integer
      month_number:
        type: integer
      year_number:
        type: integer
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount:
    properties:
      account_identifier:
        type: string
      alias:
        type: string
      company:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
      company_id:
        type: integer
      id:
        type: integer
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API REST para gestión de gastos e ingresos del hogar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: api-home-pay API
  version: "1.0"
paths:
...
  /api/expenses:
    get:
      consumes:
      - application/json
      description: Get all expenses for the authenticated user with optional filters
      parameters:
      - description: Filter by period ID
        in: query
        name: period_id
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Filter by service account ID
        in: query
        name: account_id
        type: integer
      - description: Filter by payment status (paid/pending)
        in: query
        name: payment_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Create a new expense
      parameters:
      - description: Expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create expense
      tags:
      - expenses
  /api/expenses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete expense
      tags:
      - expenses
    get:
      consumes:
      - application/json
      description: Get a specific expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get expense by ID
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update expense
      tags:
      - expenses
  /api/expenses/{id}/pay:
    patch:
      consumes:
      - application/json
      description: Mark an expense as paid
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark expense as paid
      tags:
      - expenses
  /api/expenses/pending:
    get:
      consumes:
      - application/json
      description: Get pending expenses with optional filters for days ahead and overdue
      parameters:
      - description: 'Number of days to look ahead (default: 7)'
        in: query
        name: days_ahead
        type: integer
      - description: Only return overdue expenses
        in: query
        name: overdue_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Expense'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pending expenses
      tags:
      - expenses
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"